Implement pthreads_available() by checking for pthread_create().
We currently detect whether the application we are running in was
linked against libpthread by checking the address of pthread_once()
against NULL (after defining it as a weak symbol), but this doesn't
work on DragonFly BSD, FreeBSD, or NetBSD, as on those platforms,
pthread_once() is provided by libc.  Check for pthread_create()
instead of pthread_once(), as that seems to work on all platforms.

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>